Move fixtures up a level
